refactor(reposicao): extrai tabs (Detalhes/Itens/Negociação) de AdminReposicaoPromocaoDetail
Continua a quebra do god-component. Extrai os 3 painéis de TabsContent
para src/components/reposicao/promocaoDetail/ com props explícitas:
- DetalhesTab (card de extração Vision + formulário)
- ItensTab (tabela com edição inline + linha de adição)
- NegociacaoTab (histórico de eventos)

Page: 1205 → 534 linhas. Comportamento preservado; imports mortos removidos.
tsc --noEmit limpo, 358/358 testes.
